Proper refactoring of post a job popup. (#297)
* Proper refactoring of post a job popup.

* Remove unnecessary prop.
