+++ src/Components/Dashboard/UserProfile.js
@@ -5,7 +5,7 @@ import { Button } from "@material-ui/core";
 import { faMapMarker, faBuilding } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import EscalationsCard from "./Cards/EscalationsCard";
-import { httpGet, httpPost } from "../../lib/dataAccess";
+import { httpGet } from "../../lib/dataAccess";
 import jwtDecode from "jwt-decode";
 import Dialog from "@material-ui/core/Dialog";
 // import IndustryTags from "../Registration/industry_tags";
@@ -19,7 +19,7 @@ import FAQPopup from "./Popups/FAQPopup";
 import PostJobPopup from "./Popups/PostJobPopup";
 
 const useStyles = makeStyles((theme) => ({
-  root1: {
+  root: {
     display: "flex",
     flex: "1",
     overflowX: "hidden",
@@ -51,6 +51,7 @@ const useStyles = makeStyles((theme) => ({
     textAlign: "center",
     display: "block",
   },
+
   city: {
     margin: "2vh 40px 1vh 40px",
     fontFamily: "myriad-pro, sans-serif",
@@ -69,7 +70,7 @@ const useStyles = makeStyles((theme) => ({
     fontWeight: "bold",
   },
 
-  numCoffeChat: {
+  numCoffeeChat: {
     margin: "1vh 40px 1vh 40px",
     fontFamily: "myriad-pro, sans-serif",
     fontSize: "15px",
@@ -134,7 +135,7 @@ const useStyles = makeStyles((theme) => ({
     objectFit: "cover",
   },
 
-  button: {
+  creditsButton: {
     textTransform: "none",
     backgroundColor: "#B5A165",
     marginBottom: "1vh",
@@ -148,7 +149,6 @@ const useStyles = makeStyles((theme) => ({
       backgroundColor: "#F1F1F1",
       color: "#484848",
     },
-
     fontSize: "15px",
     fontWeight: "bold",
     fontFamily: "myriad-pro, sans-serif",
@@ -156,7 +156,7 @@ const useStyles = makeStyles((theme) => ({
     paddingRight: "50px",
   },
 
-  button1: {
+  postButton: {
     textTransform: "none",
     backgroundColor: "#000000",
     marginBottom: "1vh",
@@ -177,6 +177,7 @@ const useStyles = makeStyles((theme) => ({
     paddingLeft: "75px",
     paddingRight: "75px",
   },
+
   updateProfile: {
     margin: "2vh 40px 1vh 40px",
     fontFamily: "myriad-pro, sans-serif",
@@ -186,18 +187,7 @@ const useStyles = makeStyles((theme) => ({
     color: "#6EA0B5",
     cursor: "pointer",
   },
-  contact: {
-    marginTop: "1vh",
-    marginLeft: "40px",
-    marginBottom: "1vh",
-    marginRight: "40px",
-    fontFamily: "myriad-pro, sans-serif",
-    fontSize: "12px",
-    textAlign: "center",
-    display: "block",
-    color: "#6EA0B5",
-    cursor: "pointer",
-  },
+
   faq: {
     margin: "1vh 40px 1vh 40px",
     fontFamily: "myriad-pro, sans-serif",
@@ -206,139 +196,6 @@ const useStyles = makeStyles((theme) => ({
     display: "block",
     color: "#6EA0B5",
     cursor: "pointer",
-  },
-
-  toolbar: {
-    height: "8vh",
-    backgroundColor: "black",
-    boxShadow: "0px 0px 0px",
-    width: "100%",
-  },
-
-  dialog: {
-    borderRadius: "40%",
-  },
-
-  textbox: {
-    boxShadow: "0px 0px 0px",
-    color: "white",
-    "& .MuiInput-underline:before": {
-      borderBottom: "2px solid #B6A165 ",
-    },
-    "& .MuiInput-underline:after": {
-      color: "#455E6A",
-      borderBottom: "2px solid #B6A165 ",
-    },
-    "& label.Mui-focused": {
-      color: "#455E6A",
-    },
-  },
-
-  input: {
-    boxShadow: "0px 0px ",
-  },
-
-  limit: {
-    width: "90%",
-    display: "block",
-    margin: "0 auto",
-    textAlign: "right",
-    fontSize: "10px",
-    color: "#58595B",
-  },
-
-  jobTitles: {
-    margin: "0px 0px 5px 0px",
-  },
-
-  grid: {
-    paddingLeft: "30px",
-    paddingright: "30px",
-  },
-
-  text: {
-    font: "Myriad Pro",
-  },
-
-  close: {
-    position: "absolute",
-    right: "8%",
-  },
-
-  cssOutlinedInput: {
-    "&$cssFocused $notchedOutline": {
-      borderColor: `#B6A165 `,
-    },
-  },
-
-  cssFocused: {},
-
-  notchedOutline: {
-    borderWidth: "2px",
-    borderColor: "#B6A165 ",
-  },
-
-  textField: {
-    width: "111ch",
-    margin: "5px 20px 5px 20px",
-  },
-
-  title: {
-    width: "95%",
-    display: "block",
-    margin: "15px 10px 2px 10px",
-    textAlign: "left",
-    fontSize: "16px",
-    color: "#58595B",
-  },
-
-  checkbox: {
-    color: "#B6A165",
-    "&$checked": {
-      color: "#B6A165",
-    },
-  },
-
-  checked: {},
-
-  checkboxGrid: {
-    fontSize: "7px",
-    margin: "5px 5px 0px 14px",
-  },
-
-  radioButton: {
-    color: "#58595B",
-    margin: "5px 20px 0px 30px",
-    "@media (max-width: 600px)": {
-      margin: "0px",
-      fontSize: "5px",
-    },
-    fontSize: "16px",
-  },
-
-  radioMarginFirst: {
-    margin: "15px 20px 5px 30px",
-    width: "85%",
-    "@media (max-width: 600px)": {
-      marginLeft: "0px",
-    },
-  },
-
-  radioMarginSecond: {
-    margin: "15px 20px 0px 30px",
-    width: "85%",
-    "@media (max-width: 600px)": {
-      marginLeft: "0px",
-    },
-  },
-
-  contactBox: {
-    fontSize: "7px",
-    textAlign: "center",
-    display: "block",
-    color: "#58595B",
-    cursor: "pointer",
-    marginLeft: "4px",
   },
 }));
 
@@ -358,24 +215,8 @@ class Landing extends Component {
       openFaq: false,
       active: 0,
       max_characters: 2000,
-      checkedBox: false,
       numJobs: 0,
       numChats: 0,
-      showError: false,
-      jobsData: {
-        title: "",
-        company: "",
-        country: "",
-        region: "",
-        city: "",
-        description: "",
-        requirements: "",
-        job_type: "REGULAR_JOB", // BOARD_POSITION or REGULAR_JOB
-        job_tags: [],
-        salary: 0,
-        deadline: 0,
-        can_contact: false,
-      },
       credits: jwtDecode(localStorage.getItem("idToken"))["custom:credits"],
       userType: jwtDecode(localStorage.getItem("idToken"))["custom:user_type"],
     };
@@ -406,103 +247,9 @@ class Landing extends Component {
     });
   }
 
-  onTagsChange = (event, values) => {
-    var jobsDataObj = { ...this.state.jobsData };
-    jobsDataObj.job_tags = values;
-    this.setState({
-      jobsData: jobsDataObj,
-    });
-    if (values.length > 3) {
-      this.setState({
-        showError: true,
-        errorText: "Please pick up to 3 tags",
-      });
-    } else {
-      this.setState({
-        showError: false,
-        errorText: "",
-      });
-    }
-  };
-
   postJob = (event) => {
     this.setState({
       openPostJob: true,
-      jobsData: {
-        title: "",
-        company: "",
-        country: "",
-        region: "",
-        city: "",
-        description: "",
-        requirements: "",
-        job_type: "REGULAR_JOB", // BOARD_POSITION or REGULAR_JOB
-        job_tags: [],
-        salary: 0,
-        deadline: 0,
-        can_contact: false,
-      },
-      checkedBox: false,
-    });
-  };
-
-  submitJob = async () => {
-    // check that all the required fields are set / properly set
-    if (this.state.jobsData.job_tags.length > 3) {
-      alert("There are more than 3 job tags selected.");
-      return;
-    }
-    if (
-      this.state.jobsData.title === "" ||
-      this.state.jobsData.title === undefined ||
-      this.state.jobsData.company === "" ||
-      this.state.jobsData.company === undefined ||
-      this.state.jobsData.country === "" ||
-      this.state.jobsData.country === undefined ||
-      this.state.jobsData.region === "" ||
-      this.state.jobsData.region === undefined ||
-      this.state.jobsData.city === "" ||
-      this.state.jobsData.city === undefined ||
-      this.state.jobsData.description === "" ||
-      this.state.jobsData.description === undefined ||
-      this.state.jobsData.requirements === "" ||
-      this.state.jobsData.requirements === undefined
-    ) {
-      this.props.enqueueSnackbar(
-        "One of the required fields is not set (title, company, country, region, city, description or requirements).",
-        {
-          variant: "warning",
-        }
-      );
-      return;
-    }
-
-    let idToken = (await Auth.currentSession())
-      .getIdToken()
-      .getJwtToken()
-      .toString();
-
-    // get user info and update jobs data
-    const userProfile = JSON.parse(localStorage.getItem("userProfile"));
-    var jobsDataObj = { ...this.state.jobsData };
-    jobsDataObj.posted_by = userProfile.email;
-    jobsDataObj.poster_family_name = userProfile.family_name;
-    jobsDataObj.poster_given_name = userProfile.given_name;
-
-    // post job and close popup
-    await httpPost("jobs", idToken, jobsDataObj)
-      .then((res) => {
-        this.props.enqueueSnackbar("Successfully submitted a job posting:", {
-          variant: "success",
-        });
-      })
-      .catch((err) => {
-        this.props.enqueueSnackbar("Failed:" + err, {
-          variant: "error",
-        });
-      });
-    this.setState({
-      openPostJob: false,
     });
   };
 
@@ -542,43 +289,6 @@ class Landing extends Component {
     });
   };
 
-  handleFormDataChange = (id) => (event) => {
-    var jobsDataObj = { ...this.state.jobsData };
-
-    if (id === "requirements") {
-      jobsDataObj.requirements = event.target.value
-        .toString()
-        .slice(0, this.state.max_characters);
-    } else if (id === "description") {
-      jobsDataObj.description = event.target.value
-        .toString()
-        .slice(0, this.state.max_characters);
-    } else if (id === "salary") {
-      jobsDataObj.salary = event.target.value;
-    } else if (id === "city") {
-      jobsDataObj.city = event.target.value;
-    } else if (id === "region") {
-      jobsDataObj.region = event.target.value;
-    } else if (id === "country") {
-      jobsDataObj.country = event.target.value;
-    } else if (id === "company") {
-      jobsDataObj.company = event.target.value;
-    } else if (id === "title") {
-      jobsDataObj.title = event.target.value;
-    } else if (id === "job_type") {
-      jobsDataObj.job_type = event.target.value;
-    } else if (id === "contact_me") {
-      jobsDataObj.can_contact = !this.state.jobsData.can_contact;
-      this.setState({
-        checkedBox: event.target.checked,
-      });
-    }
-
-    this.setState({
-      jobsData: jobsDataObj,
-    });
-  };
-
   getUserProfile = () => {
     const userProfile = JSON.parse(localStorage.getItem("userProfile"));
     let userLocation = "N/A";
@@ -609,8 +319,9 @@ class Landing extends Component {
   render() {
     const classes = this.props.classes;
     const userProfile = this.getUserProfile();
+
     return (
-      <div className={classes.root1}>
+      <div className={classes.root}>
         <div style={{ margin: "auto" }}>
           <img
             className={classes.image}
@@ -644,10 +355,10 @@ class Landing extends Component {
               />
               <b>{userProfile.company} </b>
             </p>
-            <p className={classes.numCoffeChat}>
+            <p className={classes.numCoffeeChat}>
               {userProfile.numCoffeeChats} Coffee Chats
             </p>
-            <p className={classes.numCoffeChat}>
+            <p className={classes.numCoffeeChat}>
               {userProfile.numJobApplications} Jobs Applied To
             </p>
             <div className={classes.circle}>
@@ -663,7 +374,7 @@ class Landing extends Component {
           {this.state.userType !== "MENTOR" ? (
             <span>
               <Button
-                className={classes.button}
+                className={classes.creditsButton}
                 variant="contained"
                 onClick={this.handleCreditsOpen}
               >
@@ -686,7 +397,7 @@ class Landing extends Component {
           {jwtDecode(localStorage.getItem("idToken"))["custom:user_type"] !==
           "FREE" ? (
             <Button
-              className={classes.button1}
+              className={classes.postButton}
               variant="contained"
               onClick={this.postJob}
             >
@@ -699,16 +410,8 @@ class Landing extends Component {
             FAQ
           </p>
           <PostJobPopup
-            jobsData={this.state.jobsData}
             openPostJob={this.state.openPostJob}
-            submitJob={this.submitJob}
             handlePostJobClose={this.handlePostJobClose}
-            handleFormDataChange={this.handleFormDataChange}
-            onTagsChange={this.onTagsChange}
-            showError={this.state.showError}
-            errorText={this.state.errorText}
-            checkedBox={this.state.checkedBox}
-            classes={classes}
           />
           <FAQPopup
             openFaq={this.state.openFaq}
